Refactor presigned URL handling; update numOfUrls type to String, enhance request processing in routes, and improve LLM inference retry logic.

+++ src/main/scala/business/GetPresignedURLSvcFlow.scala
@@ -59,7 +59,7 @@ class GetPresignedURLSvcFlow(cfgCtx: CfgCtx)
                     generatePutPresignedUrls(
                       presigner,
                       result.imageRepoId,
-                      numOfUrls.getOrElse(throw new Exception("Invalid request"))
+                      numOfUrls.getOrElse(throw new Exception("Invalid request")).toInt
                     )
 
                   case PresignedUrlType.GET =>
